correct setMovieRect()

+++ src/eSignRGBColumn.h
@@ -12,7 +12,7 @@ class eSignRGBColumn
         ofRectangle columnRect;
 
 
-        void setup(float x, float y, int LEDNumber, bool direction, float rectSize);
+        void setup(float x, float y, int LEDNumber, bool direction, int rectSize);
         void setColor(ofColor c);
         ofColor getColor();
         int portNumber;
